+++ roadgraph_builder/hd/__init__.py
@@ -3,11 +3,15 @@
 from roadgraph_builder.hd.boundaries import centerline_lane_boundaries, polyline_to_json_points
 from roadgraph_builder.hd.lidar_fusion import fuse_lane_boundaries_from_points
 from roadgraph_builder.hd.pipeline import SDToHDConfig, enrich_sd_to_hd
+from roadgraph_builder.hd.refinement import EdgeHDRefinement, apply_refinements_to_graph, refine_hd_edges
 
 __all__ = [
+    "EdgeHDRefinement",
     "SDToHDConfig",
+    "apply_refinements_to_graph",
     "centerline_lane_boundaries",
     "enrich_sd_to_hd",
     "fuse_lane_boundaries_from_points",
     "polyline_to_json_points",
+    "refine_hd_edges",
 ]